Use `unsafeGet` when getting path (#56)
Use `unsafeGet` since we already check that the path exists. There is
still an assertion inside that we can probably remove since this is a
hot path. This is just a minor edit that I saw in the github UI and was
kinda curious if there is any performance gains