Proc based API (#61)
Start of the move to making everything be based on procedures instead of
a DSL. Aim is to have a fully typed API so users don't need to use `ctx`
anymore.

Will eventually lead to a 2.0 breaking version, for now I compile the
DSL into the procedure based code. Will slowly expand on the helpers and
API as I go along, but want to start getting this in mainline to stop
bit rot